Conflicts:
	IGVC_Control/IGVC_Test.pro
	src/intelligence/linedetection/transformer.cpp
	src/intelligence/linedetection/transformer.h
	src/tests/testlinedetection.hpp
	src/tests/testpositiontracker.hpp

+++ src/common/datastructures/ImageData.hpp
@@ -10,7 +10,7 @@ class ImageData : public SensorData
   public:
     ImageData(): SensorData() {}
     ImageData(cv::Mat image): SensorData(), _img(image) {}
-    ImageData(cv::Mat image, long long int time_us): SensorData(time_us), _img(image) {}
+    ImageData(cv::Mat image, unsigned long int time_us): SensorData(time_us), _img(image) {}
     ImageData deepCopy() { return ImageData(_img.clone(), getTimeMicroSeconds()); }
     cv::Mat& mat() {return _img;}
     virtual ~ImageData() {}
